+++ src/observe.js
@@ -79,14 +79,14 @@
 
                 } else {
                     if ( !_configuration.safeMode ) {
-                        console.error( "The attribute '" + _attribute_Name_Watch_Options + "' is not a valid object." );
+                        console.error( _configuration.attributeNotValidErrorText.replace( "{{attribute_name}}", _attribute_Name_Watch_Options ) );
                         result = false;
                     }
                 }
 
             } else {
                 if ( !_configuration.safeMode ) {
-                    console.error( "The attribute '" + _attribute_Name_Watch_Options + "' has not been set correctly." );
+                    console.error( _configuration.attributeNotSetErrorText.replace( "{{attribute_name}}", _attribute_Name_Watch_Options ) );
                     result = false;
                 }
             }
@@ -438,6 +438,10 @@
         return isDefinedNumber( value ) ? value : defaultValue;
     }
 
+    function getDefaultString( value, defaultValue ) {
+        return isDefinedString( value ) ? value : defaultValue;
+    }
+
     function getDefaultDate( value, defaultValue ) {
         return isDefinedDate( value ) ? value : defaultValue;
     }
@@ -484,7 +488,7 @@
                 }
                 
             } catch ( e2 ) {
-                parsed = logError( "Errors in object: " + e1.message + ", " + e2.message );
+                parsed = logError( _configuration.objectErrorText.replace( "{{error_1}}",  e1.message ).replace( "{{error_2}}", e2.message ) );
                 result = null;
             }
         }
@@ -764,6 +768,14 @@
     function buildDefaultConfiguration() {
         _configuration.safeMode = getDefaultBoolean( _configuration.safeMode, true );
         _configuration.domElementTypes = getDefaultStringOrArray( _configuration.domElementTypes, [ "*" ] );
+
+        buildDefaultConfigurationStrings();
+    }
+
+    function buildDefaultConfigurationStrings() {
+        _configuration.objectErrorText = getDefaultString( _configuration.objectErrorText, "Errors in object: {{error_1}}, {{error_2}}" );
+        _configuration.attributeNotValidErrorText = getDefaultString( _configuration.attributeNotValidErrorText, "The attribute '{{attribute_name}}' is not a valid object." );
+        _configuration.attributeNotSetErrorText = getDefaultString( _configuration.attributeNotSetErrorText, "The attribute '{{attribute_name}}' has not been set correctly." );        
     }