+++ docs/CHANGE_LOG.md
@@ -2,6 +2,11 @@
 
 ## Version 0.8.0:
 
+#### **Configuration Options:**
+- Added a new configuration option called "objectErrorText", which states the error text that should be shown when an object error is detected (defaults to "Errors in object: {{error_1}}, {{error_2}}").
+- Added a new configuration option called "attributeNotValidErrorText", which states the error text that should be shown when a binding object is'nt valid (defaults to "The attribute '{{attribute_name}}' is not a valid object.").
+- Added a new configuration option called "attributeNotSetErrorText", which states the error text that should be shown when a binding attribute is'nt set (defaults to "The attribute '{{attribute_name}}' has not been set correctly.").
+
 #### **General Improvements:**
 - Objects are now checked and defaulted properly when invalid values are passed.